+++ src/routers/post-routers.ts
@@ -2,10 +2,12 @@ import {Request, Response, Router} from "express";
 import {postsValidation} from "../middlewares/posts/posts-validation";
 import {inputValidationMiddleware} from "../middlewares/input-validation-middleware";
 import {authorizationMiddleware} from "../middlewares/authorization";
-import {RequestQueryParams} from "../db/types/query-types";
+import {RequestParams, RequestQueryParams} from "../db/types/query-types";
 import {getPaginationData} from "../utils/pagination.utility";
 import {getSortPostsQuery} from "../utils/posts-query.utility";
 import {postsService} from "../domain/posts-service";
+import {ObjectId} from "mongodb";
+import {commentValidation} from "../middlewares/posts/comment-validation";
 
 
 
@@ -25,12 +27,47 @@ postsRouter.get('/', async (req: RequestQueryParams<{sortBy: string, sortDirecti
 })
 
 postsRouter.get('/:id', async (req: Request, res: Response) => {
+    if(!ObjectId.isValid(req.params.id)){
+        res.sendStatus(404)
+        return
+    }
     let post = await postsService.getPostById(req.params.id)
     if (post) {
         res.status(200).send(post)
     } else {
         res.sendStatus(404)
     }
+})
+
+postsRouter.get('/:id/comments', async (req: RequestParams<{id: string},{sortBy: string, sortDirection: string, pageNumber: number, pageSize: number}>, res: Response) => {
+
+    const checkPostsComments = await postsService.checkPostCommentCollection(req.params.id)
+
+    if(checkPostsComments) {
+        const postsQuery = getSortPostsQuery(req.query.sortBy, req.query.sortDirection)
+        const pagination = getPaginationData(req.query.pageNumber, req.query.pageSize);
+
+        const postCommentsList = await postsService.getAllPostsComments({
+            ...postsQuery,
+            ...pagination
+        });
+
+        res.send(postCommentsList)
+    } else {
+        res.sendStatus(404)
+        return
+    }
+
+
+    const postsQuery = getSortPostsQuery(req.query.sortBy, req.query.sortDirection)
+    const pagination = getPaginationData(req.query.pageNumber, req.query.pageSize);
+
+    const postList = await postsService.getAllPostsComments({
+        ...postsQuery,
+        ...pagination
+    });
+
+    res.send(postList)
 })
 
 postsRouter.post('/',
@@ -42,11 +79,35 @@ postsRouter.post('/',
         res.status(201).send(newPosts)
     })
 
+postsRouter.post('/:id/comments',
+    authorizationMiddleware,
+    commentValidation,
+    inputValidationMiddleware,
+    async (req: Request, res: Response) => {
+        if(!ObjectId.isValid(req.params.id)){
+            res.sendStatus(404)
+            return
+        }
+
+        const checkID = await postsService.checkPostCollection(req.params.id)
+
+        if(checkID) {
+            const newPostComment = await postsService.createNewPostComment(req.params.id, req.body)
+            res.status(201).send(newPostComment)
+        } else {
+            res.sendStatus(404)
+        }
+    })
+
 postsRouter.put('/:id',
     authorizationMiddleware,
     postsValidation,
     inputValidationMiddleware,
     async (req: Request, res: Response) => {
+        if(!ObjectId.isValid(req.params.id)){
+            res.sendStatus(404)
+            return
+        }
         const isUpdate = await postsService.updatePostById(req.params.id, req.body)
         if (isUpdate) {
             const post = await postsService.getPostById(req.params.id)
@@ -57,6 +118,10 @@ postsRouter.put('/:id',
     })
 
 postsRouter.delete('/:id', authorizationMiddleware, async (req: Request, res: Response) => {
+    if(!ObjectId.isValid(req.params.id)){
+        res.sendStatus(404)
+        return
+    }
     const isDelete = await postsService.deletePostById(req.params.id)
     if (isDelete) {
         res.sendStatus(204)
